Instead of having certain ticks where all the getters and putters are updated the mod now distributes the work over 60 ticks.
Putters split their work over 60 ticks. Getters work is splitted up into two parts, collecting requests and fulfilling requests. The first 40 ticks is used to collect requests and the last 20 are used to fulfill them. Haven't tested if this ratio is optimal but it's good enough for the most part.

The work distribution is simply implemented by using a doubly linked list with an iterator that goes over a certain amount of entities each tick. I chose a doubly linked list because it's easy to remove and add entities to it while it's being iterated over.

Putters are ~60x times faster. Tested in waterworld where the peak tick time went from ~50ms to ~0.8ms. Waterworld has 5504 put tanks so this pr should allow +50.000 putters to work smoothly on the same map.

Getters are around 4.5x faster. Tested in mustafar where the peak tick time went from ~4.5ms to ~1.0ms. Getters are more cpu intensive than putters but this pr should still make a world work smoothly with +10.000 getters.