@Danielv123 2018-05-11
Does this still work properly if the timings of imports/exports are changed, ex changing the chest polling time, or the polling timers in client.js, or sending multiple smaller import commands with a few items each instead of one massive one?